style: rename usexxParams
