+++ front/app/advanced-search/components/DownloadingButton.tsx
@@ -5,12 +5,12 @@ import Link from 'next/link';
 import { Button } from '@/components/ui/button';
 import { createAdvancedSearchDownloadingLink } from '@/utils/fetchers/advanced-search/download/downloadAdvancedSearch-client';
 
-import { useFiltersParams } from '../hooks/useFiltersFromSearchParams';
-import { usePaginationFromSearchParams } from '../hooks/usePaginationFromSearchParams';
+import { useFiltersParams } from '../hooks/useFiltersParams';
+import { usePaginationParams } from '../hooks/usePaginationParams';
 
 export default function DownloadingButton() {
   const { filters } = useFiltersParams();
-  const { pagination } = usePaginationFromSearchParams();
+  const { pagination } = usePaginationParams();
   const downloadingLink = createAdvancedSearchDownloadingLink(filters, pagination).toString();
 
   return (
