+++ front/app/advanced-search/components/Filters/SelectMarchesPublicsScore.tsx
@@ -1,6 +1,6 @@
 import { TransparencyScore } from '@/components/TransparencyScore/constants';
 
-import { useFiltersParams } from '../../hooks/useFiltersFromSearchParams';
+import { useFiltersParams } from '../../hooks/useFiltersParams';
 import { Selector } from './Selector';
 
 export function SelectMarchesPublicsScore() {
